+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/_internal/kotlinSrcSets.kt
@@ -1,5 +1,6 @@
 package com.javiersc.hubdle.project.extensions._internal
 
+import org.gradle.api.NamedDomainObjectSet
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.findByType
 import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
@@ -17,8 +18,8 @@ internal val Project.kotlinSourceSetCommonMain: KotlinSourceSet?
             it.name == "commonMain"
         }
 
-internal val Project.kotlinSourceSetMainOrCommonMain: KotlinSourceSet?
+internal val Project.kotlinSourceSetMainOrCommonMain: NamedDomainObjectSet<KotlinSourceSet>?
     get() =
-        extensions.findByType<KotlinProjectExtension>()?.sourceSets?.firstOrNull {
-            it.name == "main" || it.name == "commonMain"
+        extensions.findByType<KotlinProjectExtension>()?.sourceSets?.named { name ->
+            name == "main" || name == "commonMain"
         }